separate smart/dumb components
