+++ js/app/components/IFrame/iframe-css.css
@@ -1,0 +1,5 @@
+:local(.paymentIframe) {
+	border: 0;
+	width: 100%;
+	height: 50vh;
+}
