+++ js/app/components/InvoiceContainer/invoice-view.css
@@ -142,9 +142,3 @@ tbody > tr:last-child > td:first-child {
     align-items: center;
     flex-direction: column;
 }
-
-:local(.paymentIframe) {
-	border: 0;
-	width: 100%;
-	height: 50vh;
-}
